LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/css/cssom-view/elementFromPoint-expected.txt

 
 
    
Hello!
 
Another teal
 

PASS Negative co-ordinates 
PASS co-ordinates larger than the viewport 
PASS co-ordinates larger than the viewport from in iframe 
PASS Return first element that is the target for hit testing 
PASS First element to get mouse events with pointer-events css 
PASS SVG element at x,y 
PASS transformed element at x,y 
PASS no hit target at x,y 
PASS No viewport available 
FAIL Image Maps assert_equals: Should have returned the area element expected Element node <area id="rectG" shape="rect" coords="0,0,90,100" href="#... but got Element node <html><head><title>cssom-view - elementFromPoint</title>
...
FAIL Fieldsets assert_equals: The fieldset should not cover up the div it doesn't even overlap expected Element node <div id="fieldset-div" class="size" style="position: abso... but got null
